version 1.0, Shift select funtion for ease of selection, distribution recommendation currently based on global target. Select form distinguished from update form. Need information on forms page to display current climb

+++ climbs/views.py
@@ -8,22 +8,52 @@ from django_tables2 import RequestConfig
 from django.db.models import Count, Min, Max
 import datetime
 
-DATE = datetime.date.today()
 
-# Create your views here.
+### Globals
+
+DATE = datetime.date.today()
+CLIMBTYPE = {'route': 2, 'boulder': 1}
+
 def get_user(request):
     pk = request.user.pk
     setter = get_object_or_404(Setter, user__pk=pk)
     return setter
+def is_foreman(setter):
+    user = User.objects.get(username = setter.user.username)
+    return user.groups.filter(name='Foreman').exists()
 
-CLIMBTYPE = {'route': 2, 'boulder': 1}
-GYM = 'Pipeworks'
 
-
+### Views
 
 def climb_query(request, climb_type='boulder'):
     gym = get_user(request).current_gym
-    query = Climb.objects.values('date_created', 'area__location_name').filter(area__gym__name=gym).order_by('date_created').annotate(count = Count('grade'), min_grade=Min('grade__pk'), max_grade=Max('grade__pk'))
+    try:
+        climb_selection = request.session['select_climbs']
+    except:
+        climb_selection =[]
+
+    if request.method == "POST" and request.session.get('round_two'):
+        climb_ids = list(request.POST.getlist('selection'))
+        request.session['remove_climbs'] = climb_ids
+        num_climbs = len(climb_ids)
+        climbs = Climb.objects.filter(pk__in=climb_ids)
+        areas = Climb.objects.values('area__location_name').filter(pk__in=climb_ids).annotate(count=Count('grade'))
+        table = ClimbRemoveTable(climbs)
+        return render(request, 'climbs/replace_climbs.html', {'table': table, 'areas': areas, 'gym':gym, 'climb_type':climb_type })
+
+    elif request.method =="POST":
+        request.session['round_two'] = True
+        date_created = list(request.POST.getlist('selection'))
+        climb_selection = date_created
+        climbs = Climb.objects.filter(date_created__in=date_created, status__status='current')
+        areas = Climb.objects.values('area__location_name').filter(date_created__in=date_created).annotate(count=Count('grade'))
+        table = ClimbQueryTable(climbs)
+        area = Area.objects.get(location_name='Alcove')
+        return render(request, 'climbs/climbs_list.html', {'table': table, 'areas': areas, 'gym':gym, 'climb_type':climb_type })
+
+    #Present summary of climbs by date
+    query = Climb.objects.values('date_created', 'area__location_name').filter(area__gym__name=gym, status__status='current').order_by('date_created').annotate(count = Count('grade'), min_grade=Min('grade__pk'), max_grade=Max('grade__pk'))
+
     table_data = []
     for item in list(query):
         grade_pk = []
@@ -49,47 +79,32 @@ def climb_query(request, climb_type='boulder'):
     return render(request, 'climbs/climbs_list.html', {'table': table, 'gym': gym})
 
 
-###TODO: filter based on climb type to provide the correct grade type
-def distribution(reset_num, climb_type= CLIMBTYPE['boulder'], gym = GYM):
-    """
-    returns list of tuples for the difference between desired climbs for grade and
-    actual climbs per grade. Spread is gym scope not area scope.
-    """
-    test = Spread.objects.values('grade__grade', 'quantity').filter(gym__name=GYM, grade__climb=climb_type)
-    climbs = Climb.objects.values('grade__grade').annotate(total=Count('grade')).filter(area__gym__name=GYM, grade__climb=climb_type, status__id = 1)
-    grade_spread = []
-    needed = 0
-    for climb in climbs:
-        climb_grade = climb['grade__grade']
-        for desired in test:
-            spread_grade = desired['grade__grade']
-            if climb_grade == spread_grade:
-
-                diff = int(desired['quantity']) - int(climb['total'])
-                grade_spread.append((climb['grade__grade'], diff))
-                needed = needed + diff
-    finalSpread = []
-    for item in grade_spread:
-        if item[1] > 0:
-            per_grade = round(float(item[1])/float(needed) * float(reset_num))
-            if per_grade >0:
-                finalSpread.append([item[0], int(per_grade)])
-    return finalSpread
-
 def climb_verification(request):
     """
     take a queryset from the days climbs and populate in a formset.
     upon verification update all climbs from 'in progress' to 'current'.
     """
-    gym = get_user(request).current_gym
-    queue = Climb.objects.all().filter(status=4, area__gym__name = gym).delete()
-    in_progress = Climb.objects.all().filter(status=3, area__gym__name = gym)
+    setter = get_user(request)
+    gym = setter.current_gym
 
-    status = Status.objects.get(status='current')
-    for climb in in_progress:
-        climb.status = status
-        climb.save()
-    return redirect('home')
+    if request.method == "POST":
+
+        gym = get_user(request).current_gym
+        queue = Climb.objects.all().filter(status=4, area__gym__name = gym).delete()
+        in_progress = Climb.objects.all().filter(status=3, area__gym__name = gym)
+
+        status = Status.objects.get(status='current')
+        for climb in in_progress:
+            climb.status = status
+            climb.save()
+        return redirect('home')
+
+
+    in_progress = Climb.objects.all().filter(status=3, area__gym__name = gym)
+    in_progress_table = ClimbTable(in_progress)
+    RequestConfig(request).configure(in_progress_table)
+
+    return render(request, 'climbs/verification.html', {'table': in_progress_table})
 
 def climbs_list(request, climb_type, template_name='climbs/climbs_list.html'):
     gym = get_user(request).current_gym
@@ -98,14 +113,16 @@ def climbs_list(request, climb_type, template_name='climbs/climbs_list.html'):
         request.session['remove_climbs'] = climb_ids
         num_climbs = len(climb_ids)
         climbs = Climb.objects.filter(pk__in=climb_ids)
+        areas = Climb.objects.values('area__location_name').filter(pk__in=climb_ids).annotate(count=Count('grade'))
         table = ClimbRemoveTable(climbs)
-
-        return render(request, 'climbs/replace_climbs.html', {'table': table, 'num_climbs': num_climbs, 'gym':gym, 'climb_type':climb_type })
+        area = Area.objects.get(location_name='Alcove')
+        return render(request, 'climbs/replace_climbs.html', {'table': table, 'areas': areas, 'gym':gym, 'climb_type':climb_type })
 
     climbs = Climb.objects.filter(area__gym__name = gym).filter(grade__climb=CLIMBTYPE[climb_type], status__status='current')
     table = ClimbTable(climbs)
     RequestConfig(request).configure(table)
     return render(request, template_name, {'table': table, 'gym': gym})
+
 
 def remove_climb(request, pk):
     Climb.objects.get(pk=pk).delete()
@@ -117,25 +134,46 @@ def climb_create(request, template_name='climbs/climb_form.html'):
         form = ClimbCreateForm(request.POST)
         if form.is_valid():
             climb = form.save(commit=False)
-            climb.status = Status.objects.get(status='in progress')
+            climb.status = Status.objects.get(status='in queue')
             climb.save()
             return redirect('climb_set')
     else:
         form = ClimbCreateForm()
     return render(request, template_name, {'form': form})
 
-def climb_update(request, pk, template_name='climbs/climb_form.html'):
+
+def climb_update(request, pk):
     climb = Climb.objects.get(pk=pk)
 
-    if request.method == 'POST':
-        form = ClimbForm(request.POST)
-        setter = get_user(request)
-
+    if request.method == "POST":
+        form = ForemanClimbUpdateForm(request.POST)
         if form.is_valid():
             update = form
             climb.color = update.cleaned_data['color']
-            climb.area = update.cleaned_data['area']
+            if update.cleaned_data['anchor']:
+                climb.anchor = update.cleaned_data['anchor']
+            # climb.status = Status.objects.get(pk=3)
+            climb.setter = update.cleaned_data['setter']
+            climb.date_created = update.cleaned_data['date_created']
             climb.grade = update.cleaned_data['grade']
+            climb.area = update.cleaned_data['area']
+            climb.save()
+            return redirect('climb_set')
+
+    else:
+        form = ForemanClimbUpdateForm(instance = climb)
+    return render(request, 'climbs/climb_form.html', {'form': form})
+
+
+def climb_select(request, pk, template_name='climbs/climb_form.html'):
+    climb = Climb.objects.get(pk=pk)
+    setter = get_user(request)
+
+    if request.method == 'POST'and is_foreman(setter):
+        form = ForemanClimbSelectForm(request.POST)
+        if form.is_valid():
+            update = form
+            climb.color = update.cleaned_data['color']
             if update.cleaned_data['anchor']:
                 climb.anchor = update.cleaned_data['anchor']
             climb.status = Status.objects.get(pk=3)
@@ -143,37 +181,61 @@ def climb_update(request, pk, template_name='climbs/climb_form.html'):
 
             climb.save()
             return redirect('climb_set')
+
+
+    elif request.method == "POST":
+        form = CimbSelectForm(request.POST)
+        if form.is_valid():
+            update = form
+            climb.color = update.cleaned_data['color']
+            if update.cleaned_data['anchor']:
+                climb.anchor = update.cleaned_data['anchor']
+            climb.status = Status.objects.get(pk=3)
+            climb.setter = setter
+
+            climb.save()
+            return redirect('climb_set')
+
+
     else:
-        form = ClimbForm(instance= climb)
+        if is_foreman(setter):
+            form = ForemanClimbSelectForm(instance = climb)
+        else:
+            form = ClimbSelectForm(instance= climb)
     return render(request, template_name, {'form': form})
 
-def formset(request, template_name='climbs/addmany_form.html'):
-    if request.method == 'POST':
-        form = AddmanyFormset()
-    else:
-        form=AddmanyFormset()
-    return render(request, template_name, {'formset': form})
+# def formset(request, template_name='climbs/addmany_form.html'):
+#     if request.method == 'POST':
+#         form = AddmanyFormset()
+#     else:
+#         form=AddmanyFormset()
+#     return render(request, template_name, {'formset': form})
 
 
 def climb_set(request):
     setter = get_user(request)
     gym = setter.current_gym
+    request.session['round_two'] = False
     try:
         climb_ids = request.session['remove_climbs']
     except:
         climb_ids =[]
 
     if request.method == 'POST' and len(climb_ids) > 0:
+        # Retire Climbs
         Climb.objects.all().filter(pk__in=climb_ids).update(status=2, date_retired=DATE)
-        set_num = request.POST.get('quantity', '')
+        # Retrieve areas for new climbs
+        areas = Climb.objects.values('area__location_name').filter(pk__in=climb_ids).annotate(count=Count('grade'))
         climb_type = Climb.objects.get(id=climb_ids[0]).grade.climb
-        table_values = distribution(set_num, gym=gym, climb_type=climb_type)
-        grade_spread = []
-        for row in table_values:
-            grade = [row[0]]
-            grade_spread +=grade * row[1]
+        request.session['remove_climbs'] = ''
 
-        Climb.objects.bulk_create([Climb(date_created=DATE, color=Color.objects.get(pk=1), grade=Grade.objects.get(grade=grade), status=Status.objects.get(id=4), setter=Setter.objects.get(id=5), area=Area.objects.get(id=1))for grade in grade_spread])
+        for area in areas:
+            print(area)
+            set_num = int(request.POST.get(area['area__location_name']))
+            area_obj = Area.objects.get(location_name=area['area__location_name'], gym=gym)
+            grades = area_obj.get_climbs_to_set(set_num, climb_type)
+
+            Climb.objects.bulk_create([Climb(date_created=DATE, color=Color.objects.get(pk=1), grade=Grade.objects.get(grade=grade), status=Status.objects.get(id=4), setter=Setter.objects.get(id=5), area=area_obj)for grade in grades])
 
         queue = Climb.objects.all().filter(status=4, area__gym__name = gym)
         queue_table = QueueTable(queue)
